requirements.txt
fastapi
jwt
pydantic
python-dotenv